fixes: cast issues
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/soma/soma-io/trunk@76608 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/libcartobase/type/voxelrgb_def.h
@@ -570,7 +570,7 @@ namespace carto {
       std::cout << "RGB:: RGB - double" << std::endl;
     #endif
     VoxelRGB result( aa );
-    return result -= bb;
+    return result -= static_cast<uint8_t>(bb);
   }
   
   inline
@@ -580,7 +580,7 @@ namespace carto {
       std::cout << "RGB:: uint8_t - RGB" << std::endl;
     #endif
     VoxelRGB result( aa );
-    return result -= static_cast<uint8_t>(bb);
+    return result -= bb;
   }
   
   inline
